+++ source/files/fileGriddedDataTimeSeries.cpp
@@ -126,11 +126,12 @@ Matrix InFileGriddedDataTimeSeries::data(const Time &time)
 {
   try
   {
-    if((time < times_.front()) || (time >= times_.back()))
-      throw(Exception(time.dateTimeStr()+" outside interval ["+times_.front()%"%D_%T, "s+times_.back()%"%D_%T) of <"s+file.fileName().str()+">"));
+    if((time < times_.front()) || (time > times_.back()))
+      throw(Exception(time.dateTimeStr()+" outside interval ["+times_.front()%"%D_%T, "s+times_.back()%"%D_%T] of <"s+file.fileName().str()+">"));
 
     // find time interval and read missing nodes
-    const UInt idx   = std::distance(times_.begin(), std::upper_bound(times_.begin(), times_.end(), time))-1;
+    const UInt idx   = std::min(std::distance(times_.begin(), std::upper_bound(times_.begin(), times_.end(), time)),
+                                static_cast<std::vector<Time>::difference_type>(times_.size()-1))-1;
     const UInt start = (idx > indexData) ? std::max(idx, indexData+data_.size()) : idx;
     const UInt end   = (idx > indexData) ? idx+data_.size() : std::min(idx+data_.size(), indexData);
     for(UInt i=start; i<end; i++)